object store: display space usage, filters
Display size and free space of the volume associated with an object
store in the object store view. This makes it easy for users to view the
amount of free space in an object store.

Fix filters in object store view: Allow object stores to be filtered by
their state and endpoints in addition to their name. This makes it easy
for users to find stopped object stores if they have too many to display
on a single page. It also makes it easy to find the object store
associated with a domain name, which otherwise would be hard if the name
of an object store has no relation to the domain name in use.

Signed-off-by: Moritz Röhrich <moritz.rohrich@suse.com>